requirements.txt
google-generativeai
rich
prompt_toolkit
requests
python-dotenv
beautifulsoup4
